Refactor People API by removing legacy GraphQL queries and consolidating service logic
- Deleted `new_people_queries.graphql` as it contained outdated GraphQL queries.
- Updated `people_router.py` to utilize the new `PeopleService`, enhancing data fetching logic.
- Removed `people_directus_service.py` to streamline service architecture and improve maintainability.
- Enhanced `people_service.py` with comprehensive methods for fetching and managing people data, ensuring better integration with the CMS.
