+++ api/src/v1/people/services/__init__.py
@@ -1,0 +1,1 @@
+"""People services package"""
